requirements.txt
numpy
pandas
matplotlib
scikit-learn
tensorflow
python-chess
